+++ spec/whois/record/parser/responses/whois.centralnic.com/br.com/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   end
   describe "#status" do
     it do
-      @parser.status.should == :registered
+      @parser.status.should == ["TRANSFER PROHIBITED"]
     end
   end
   describe "#available?" do
@@ -49,18 +49,19 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   describe "#created_on" do
     it do
       @parser.created_on.should be_a(Time)
-      @parser.created_on.should == Time.parse("2009-04-17")
+      @parser.created_on.should == Time.parse("2009-04-17 12:22:49 UTC")
     end
   end
   describe "#updated_on" do
     it do
-      lambda { @parser.updated_on }.should raise_error(Whois::PropertyNotSupported)
+      @parser.updated_on.should be_a(Time)
+      @parser.updated_on.should == Time.parse("2011-04-16 00:16:02 UTC")
     end
   end
   describe "#expires_on" do
     it do
       @parser.expires_on.should be_a(Time)
-      @parser.expires_on.should == Time.parse("2011-04-17")
+      @parser.expires_on.should == Time.parse("2014-04-17 23:59:59 UTC")
     end
   end
   describe "#nameservers" do
@@ -68,9 +69,9 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
       @parser.nameservers.should be_a(Array)
       @parser.nameservers.should have(2).items
       @parser.nameservers[0].should be_a(_nameserver)
-      @parser.nameservers[0].name.should == "ns1.terra.com.br"
+      @parser.nameservers[0].name.should == "ns1.locaweb.com.br"
       @parser.nameservers[1].should be_a(_nameserver)
-      @parser.nameservers[1].name.should == "ns2.terra.com.br"
+      @parser.nameservers[1].name.should == "ns2.locaweb.com.br"
     end
   end
 end