+++ spec/whois/record/parser/responses/whois.centralnic.com/de.com/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   end
   describe "#status" do
     it do
-      @parser.status.should == :registered
+      @parser.status.should == ["TRANSFER PROHIBITED"]
     end
   end
   describe "#available?" do
@@ -49,28 +49,31 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   describe "#created_on" do
     it do
       @parser.created_on.should be_a(Time)
-      @parser.created_on.should == Time.parse("2010-03-02")
+      @parser.created_on.should == Time.parse("2011-11-28 21:38:30 UTC")
     end
   end
   describe "#updated_on" do
     it do
-      lambda { @parser.updated_on }.should raise_error(Whois::PropertyNotSupported)
+      @parser.updated_on.should be_a(Time)
+      @parser.updated_on.should == Time.parse("2011-12-03 00:15:47 UTC")
     end
   end
   describe "#expires_on" do
     it do
       @parser.expires_on.should be_a(Time)
-      @parser.expires_on.should == Time.parse("2011-03-02")
+      @parser.expires_on.should == Time.parse("2012-11-28 23:59:59 UTC")
     end
   end
   describe "#nameservers" do
     it do
       @parser.nameservers.should be_a(Array)
-      @parser.nameservers.should have(2).items
+      @parser.nameservers.should have(3).items
       @parser.nameservers[0].should be_a(_nameserver)
-      @parser.nameservers[0].name.should == "ns.udagdns.de"
+      @parser.nameservers[0].name.should == "b.dns.gandi.net"
       @parser.nameservers[1].should be_a(_nameserver)
-      @parser.nameservers[1].name.should == "ns.udagdns.net"
+      @parser.nameservers[1].name.should == "c.dns.gandi.net"
+      @parser.nameservers[2].should be_a(_nameserver)
+      @parser.nameservers[2].name.should == "a.dns.gandi.net"
     end
   end
 end