+++ spec/whois/record/parser/responses/whois.centralnic.com/gb.net/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   end
   describe "#status" do
     it do
-      @parser.status.should == :registered
+      @parser.status.should == ["PENDING DELETE", "PENDING DELETE RESTORABLE"]
     end
   end
   describe "#available?" do
@@ -49,18 +49,19 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   describe "#created_on" do
     it do
       @parser.created_on.should be_a(Time)
-      @parser.created_on.should == Time.parse("2008-12-04")
+      @parser.created_on.should == Time.parse("2008-12-04 05:56:57 UTC")
     end
   end
   describe "#updated_on" do
     it do
-      lambda { @parser.updated_on }.should raise_error(Whois::PropertyNotSupported)
+      @parser.updated_on.should be_a(Time)
+      @parser.updated_on.should == Time.parse("2011-12-04 12:00:12 UTC")
     end
   end
   describe "#expires_on" do
     it do
       @parser.expires_on.should be_a(Time)
-      @parser.expires_on.should == Time.parse("2010-12-04")
+      @parser.expires_on.should == Time.parse("2011-12-04 23:59:59 UTC")
     end
   end
   describe "#nameservers" do
@@ -68,13 +69,13 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
       @parser.nameservers.should be_a(Array)
       @parser.nameservers.should have(4).items
       @parser.nameservers[0].should be_a(_nameserver)
-      @parser.nameservers[0].name.should == "a.ns14.net"
+      @parser.nameservers[0].name.should == "d.ns14.net"
       @parser.nameservers[1].should be_a(_nameserver)
-      @parser.nameservers[1].name.should == "b.ns14.net"
+      @parser.nameservers[1].name.should == "a.ns14.net"
       @parser.nameservers[2].should be_a(_nameserver)
       @parser.nameservers[2].name.should == "c.ns14.net"
       @parser.nameservers[3].should be_a(_nameserver)
-      @parser.nameservers[3].name.should == "d.ns14.net"
+      @parser.nameservers[3].name.should == "b.ns14.net"
     end
   end
 end