+++ spec/whois/record/parser/responses/whois.centralnic.com/kr.com/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   end
   describe "#status" do
     it do
-      @parser.status.should == :registered
+      @parser.status.should == ["TRANSFER PROHIBITED"]
     end
   end
   describe "#available?" do
@@ -49,28 +49,35 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   describe "#created_on" do
     it do
       @parser.created_on.should be_a(Time)
-      @parser.created_on.should == Time.parse("2008-06-11")
+      @parser.created_on.should == Time.parse("2008-06-11 21:25:43 UTC")
     end
   end
   describe "#updated_on" do
     it do
-      lambda { @parser.updated_on }.should raise_error(Whois::PropertyNotSupported)
+      @parser.updated_on.should be_a(Time)
+      @parser.updated_on.should == Time.parse("2011-04-17 00:17:42 UTC")
     end
   end
   describe "#expires_on" do
     it do
       @parser.expires_on.should be_a(Time)
-      @parser.expires_on.should == Time.parse("2011-06-11")
+      @parser.expires_on.should == Time.parse("2014-06-11 23:59:59 UTC")
     end
   end
   describe "#nameservers" do
     it do
       @parser.nameservers.should be_a(Array)
-      @parser.nameservers.should have(2).items
+      @parser.nameservers.should have(5).items
       @parser.nameservers[0].should be_a(_nameserver)
-      @parser.nameservers[0].name.should == "ns1.academyart.edu"
+      @parser.nameservers[0].name.should == "dbru.br.ns.els-gms.att.net"
       @parser.nameservers[1].should be_a(_nameserver)
-      @parser.nameservers[1].name.should == "ns2.academyart.edu"
+      @parser.nameservers[1].name.should == "dmtu.mt.ns.els-gms.att.net"
+      @parser.nameservers[2].should be_a(_nameserver)
+      @parser.nameservers[2].name.should == "cbru.br.ns.els-gms.att.net"
+      @parser.nameservers[3].should be_a(_nameserver)
+      @parser.nameservers[3].name.should == "cmtu.mt.ns.els-gms.att.net"
+      @parser.nameservers[4].should be_a(_nameserver)
+      @parser.nameservers[4].name.should == "ns1.academyart.edu"
     end
   end
 end