+++ spec/whois/record/parser/responses/whois.centralnic.com/uk.net/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   end
   describe "#status" do
     it do
-      @parser.status.should == :registered
+      @parser.status.should == ["OK"]
     end
   end
   describe "#available?" do
@@ -49,18 +49,19 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   describe "#created_on" do
     it do
       @parser.created_on.should be_a(Time)
-      @parser.created_on.should == Time.parse("2006-02-28")
+      @parser.created_on.should == Time.parse("2006-02-28 12:17:01 UTC")
     end
   end
   describe "#updated_on" do
     it do
-      lambda { @parser.updated_on }.should raise_error(Whois::PropertyNotSupported)
+      @parser.updated_on.should be_a(Time)
+      @parser.updated_on.should == Time.parse("2010-05-10 10:21:35 UTC")
     end
   end
   describe "#expires_on" do
     it do
       @parser.expires_on.should be_a(Time)
-      @parser.expires_on.should == Time.parse("2012-02-28")
+      @parser.expires_on.should == Time.parse("2012-02-28 23:59:59 UTC")
     end
   end
   describe "#nameservers" do
@@ -68,9 +69,9 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
       @parser.nameservers.should be_a(Array)
       @parser.nameservers.should have(2).items
       @parser.nameservers[0].should be_a(_nameserver)
-      @parser.nameservers[0].name.should == "ns1.myhostcp.com"
+      @parser.nameservers[0].name.should == "ns2.myhostcp.com"
       @parser.nameservers[1].should be_a(_nameserver)
-      @parser.nameservers[1].name.should == "ns2.myhostcp.com"
+      @parser.nameservers[1].name.should == "ns1.myhostcp.com"
     end
   end
 end