+++ spec/whois/record/parser/responses/whois.nic.la/status_available_spec.rb
@@ -21,6 +21,16 @@ describe Whois::Record::Parser::WhoisNicLa, "status_available.expected" do
     @parser = klass.new(part)
   end
 
+  describe "#referral_whois" do
+    it do
+      lambda { @parser.referral_whois }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
+  describe "#referral_url" do
+    it do
+      lambda { @parser.referral_url }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
   describe "#status" do
     it do
       @parser.status.should == []